requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
jwcrypto==1.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
wrapt==1.15.0
